refactor: separate out `WindowExt` trait into os-specific traits
