+++ packages/desktop/src/commands.rs
@@ -2,8 +2,11 @@ use std::{collections::HashMap, path::PathBuf, sync::Arc};
 
 use tauri::{State, Window};
 
+#[cfg(target_os = "macos")]
+use crate::common::WindowExtMacOs;
+#[cfg(target_os = "windows")]
+use crate::common::WindowExtWindows;
 use crate::{
-  common::WindowExt,
   config::{Config, WidgetConfig, WidgetPlacement},
   providers::{ProviderConfig, ProviderManager},
   widget_factory::{WidgetFactory, WidgetOpenOptions, WidgetState},
