+++ packages/desktop/src/common/windows/app_bar.rs
