+++ packages/desktop/src/common/windows/mod.rs
@@ -1,0 +1,5 @@
+mod app_bar;
+mod window_ext_windows;
+
+pub use app_bar::*;
+pub use window_ext_windows::*;
